requirements.txt
watchdog
python-dotenv
sqlalchemy
pandas
scikit-learn
numpy
tensorflow
joblib
twilio
requests
fastapi
uvicorn[standard]
python-multipart
lightgbm
networkx
psycopg2-binary
